+++ src/unraid_api/__init__.py
@@ -2,12 +2,17 @@
 
 from unraid_api._version import __version__
 from unraid_api.client import UnraidClient
+from unraid_api.const import (
+    MIN_API_VERSION,
+    MIN_UNRAID_VERSION,
+)
 from unraid_api.exceptions import (
     UnraidAPIError,
     UnraidAuthenticationError,
     UnraidConnectionError,
     UnraidSSLError,
     UnraidTimeoutError,
+    UnraidVersionError,
 )
 from unraid_api.models import (
     ApiKey,
@@ -33,6 +38,7 @@ from unraid_api.models import (
     NotificationOverview,
     Owner,
     ParityCheck,
+    ParityHistoryEntry,
     Permission,
     PhysicalDisk,
     Plugin,
@@ -49,10 +55,14 @@ from unraid_api.models import (
     UPSDevice,
     UserAccount,
     Vars,
+    VersionInfo,
     VmDomain,
+    format_bytes,
 )
 
 __all__ = [
+    "MIN_API_VERSION",
+    "MIN_UNRAID_VERSION",
     "ApiKey",
     "ApiKeyResponse",
     "ArrayCapacity",
@@ -76,6 +86,7 @@ __all__ = [
     "NotificationOverview",
     "Owner",
     "ParityCheck",
+    "ParityHistoryEntry",
     "Permission",
     "PhysicalDisk",
     "Plugin",
@@ -96,8 +107,11 @@ __all__ = [
     "UnraidConnectionError",
     "UnraidSSLError",
     "UnraidTimeoutError",
+    "UnraidVersionError",
     "UserAccount",
     "Vars",
+    "VersionInfo",
     "VmDomain",
     "__version__",
+    "format_bytes",
 ]